Moved PostsAppContext to dedicated provider file
Ref https://linear.app/ghost/issue/BER-2968/posts-analytics-wont-load

Extracted PostsAppContext from App.tsx to a dedicated file in the providers
folder, like PostAnalyticsContext, so that we can use it in the admin shell.
