+++ apps/posts/src/App.tsx
@@ -1,22 +1,9 @@
+import PostsAppContextProvider, {PostsAppContextType} from './providers/PostsAppContext';
 import PostsErrorBoundary from './components/errors/PostsErrorBoundary';
-import React, {createContext, useContext} from 'react';
+import React from 'react';
 import {APP_ROUTE_PREFIX, routes} from '@src/routes';
-import {AppContextType, BaseAppProps, FrameworkProvider, Outlet, RouterProvider} from '@tryghost/admin-x-framework';
+import {BaseAppProps, FrameworkProvider, Outlet, RouterProvider} from '@tryghost/admin-x-framework';
 import {ShadeApp} from '@tryghost/shade';
-
-interface PostsAppContextType extends AppContextType {
-    fromAnalytics: boolean;
-}
-
-const PostsAppContext = createContext<PostsAppContextType | undefined>(undefined);
-
-export const useAppContext = () => {
-    const context = useContext(PostsAppContext);
-    if (context === undefined) {
-        throw new Error('useAppContext must be used within an AppProvider');
-    }
-    return context;
-};
 
 interface AppProps extends BaseAppProps {
     fromAnalytics?: boolean;
@@ -40,7 +27,7 @@ const App: React.FC<AppProps> = ({framework, designSystem, fromAnalytics = false
                 refetchOnWindowFocus: false // Disable window focus refetch (Ember admin doesn't have this)
             }}
         >
-            <PostsAppContext.Provider value={appContextValue}>
+            <PostsAppContextProvider value={appContextValue}>
                 <RouterProvider prefix={APP_ROUTE_PREFIX} routes={routes}>
                     <PostsErrorBoundary>
                         <ShadeApp className="shade-posts" darkMode={designSystem.darkMode} fetchKoenigLexical={null}>
@@ -48,7 +35,7 @@ const App: React.FC<AppProps> = ({framework, designSystem, fromAnalytics = false
                         </ShadeApp>
                     </PostsErrorBoundary>
                 </RouterProvider>
-            </PostsAppContext.Provider>
+            </PostsAppContextProvider>
         </FrameworkProvider>
     );
 };
